Move non-browser lib into action
